Remove usize.

+++ packages/outpost/src/helpers.rs
@@ -46,7 +46,7 @@ pub fn decimal_param_le_one(param_value: Decimal, param_name: &str) -> Result<()
     }
 }
 
-pub fn integer_param_gt_zero(param_value: usize, param_name: &str) -> Result<(), MarsError> {
+pub fn integer_param_gt_zero(param_value: u64, param_name: &str) -> Result<(), MarsError> {
     if !param_value.gt(&0) {
         Err(MarsError::InvalidParam {
             param_name: param_name.to_string(),
